+++ packages/create-app/index.js
@@ -9,7 +9,6 @@ import ora from 'ora';
 import prompts from 'prompts';
 import { fileURLToPath } from 'url';
 import { argv, cd, chalk, fs, path, YAML } from 'zx';
-
 import { echoBrand, echoDocument } from './lib/arcblock.js';
 import { getBlockletDidList } from './lib/did.js';
 import { initGitRepo } from './lib/git.js';
@@ -27,7 +26,7 @@ import {
   toValidPackageName,
 } from './lib/utils.js';
 
-const { yellow, red, green, cyan, blue, bold } = chalk;
+const { yellow, red, green, cyan, blue, bold, magenta } = chalk;
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
@@ -143,7 +142,20 @@ const templates = [
     display: '[api] nestjs',
     color: blue,
   },
+  // dev
+  {
+    name: 'component-studio',
+    display: '[dev] component studio (beta): Local studio using for component development',
+    color: magenta,
+    // use this permanent did as blocklet.yml did always
+    permanentDid: 'z2qa7BQdkEb3TwYyEYC1psK6uvmGnHSUHt5RM',
+  },
 ];
+
+const templatesMap = {};
+templates.forEach((template) => {
+  templatesMap[template.name] = template;
+});
 
 // see: https://github.com/npm/npm/issues/3763
 const renameFiles = {
@@ -232,7 +244,9 @@ async function init() {
 
   const targetDir = argv._[0] ? String(argv._[0]) : undefined;
   const inputTemplateName = argv.template;
+
   const connectUrl = argv?.connectUrl;
+  // inputDid can be user input or command line input or getting from templates item
   const inputDid = argv.did;
   const checkRes = checkDid(inputDid);
   if (typeof checkRes === 'string') {
@@ -403,11 +417,38 @@ async function init() {
   }
 
   let didList = [];
-  if (inputDid && templateNames.length === 1) {
-    didList = [inputDid];
+  if (templateNames.length === 1 && (inputDid || templatesMap[templateNames[0]]?.permanentDid)) {
+    // permanentDid is higher priority than inputDid
+    didList = [templatesMap[templateNames[0]]?.permanentDid || inputDid];
   } else {
+    // initialize didList with null
+    didList = new Array(templateNames.length).fill(null);
+
+    const needDidTemplateNames = [];
+
+    templateNames.forEach((templateName, index) => {
+      const { permanentDid } = templatesMap[templateName];
+      if (permanentDid) {
+        didList[index] = permanentDid;
+      } else {
+        needDidTemplateNames.push(templateName);
+      }
+    });
+
     try {
-      didList = await getBlockletDidList(templateNames, connectUrl);
+      const tempDidList = await getBlockletDidList(needDidTemplateNames, connectUrl);
+      let tempIndex = 0;
+      // fill didList null item
+      // for example:
+      // templateNames = ['template1', 'template2（with permanentDid）', 'template3']
+      // didList = [null, did2, null]
+      // tempDidList = ['did1', 'did3']
+      // then didList = ['did1', 'did2', 'did3']
+      didList.forEach((did, index) => {
+        if (!did) {
+          didList[index] = tempDidList[tempIndex++];
+        }
+      });
     } catch (err) {
       console.error(red(err.message));
       process.exit(1);