+++ packages/create-app/templates/component-studio/scripts/build-clean.mjs
@@ -1,0 +1,6 @@
+/* eslint-disable no-console */
+import { rimrafSync } from 'rimraf';
+
+console.log('clean .blocklet folder');
+rimrafSync('.blocklet');
+console.log('clean .blocklet folder done!');